This is a majorly important update.
It may not seem like it when checking out the code changes, but this update houses a huge fix for Windows based hosts. It is pretty well known that Windows paths cannot exceed ~260 characters, and this offers a huge problem for Node developers because of how npm handles nested modules. I have spent approximately 15 hours today (that is not an exaggeration), but I have finally built a solution to this problem.

After many, many hours of searching, I finally came across some npm/node/vagrant issues that talk about a particular error involving Windows hosts for vagrant machines that house Node code. It took a bit of time just to narrow down that the error I was receiving was caused by this problem, but I finally did. The official NPM, Node, and Vagrant issue threads did not have any viable solutions for fixing this problem in an automated sense, so I spent the rest of my day (now ending at about 4:00 AM), and I built a solution in Vagrant that overcomes this issue.

Feel free to check out this code. I'm sorry if my commit message is hard to follow, but I am tired. Good luck, and happy coding!